[#97] refactor: 채팅방에서 업로드된 파일을 조회하는 api 파일 서버로 변경
