+++ chatroom/src/main/java/com/devcamp/flametalk/chatroom/domain/file/domain/FileRepository.java
@@ -1,7 +1,5 @@
 package com.devcamp.flametalk.chatroom.domain.file.domain;
 
-import com.devcamp.flametalk.chatroom.domain.chatroom.domain.Chatroom;
-import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 /**
@@ -9,5 +7,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
  */
 public interface FileRepository extends JpaRepository<File, Long> {
 
-  List<File> findAllByChatroom(Chatroom chatroom);
 }
