+++ file/src/main/java/com/devcamp/flametalk/domain/FileResponseMessage.java
@@ -12,6 +12,7 @@ public enum FileResponseMessage {
   FILE_CREATE_SUCCESS("파일 생성 성공"),
   FILE_DELETE_SUCCESS("파일 삭제 성공"),
   FILE_DETAIL_SUCCESS("파일 조회 성공"),
+  CHATROOM_FILES_SUCCESS("채팅방 파일 리스트 조회 성공"),
 
   FILE_CAPACITY_FAIL("파일 용량 초과");
